+++ CHANGELOG.md
@@ -17,13 +17,14 @@ project uses [Semantic Versioning](https://semver.org/).
 - Make improvements to logging-related portion of the templates
 - Add Python 3.14, drop 3.9 and 3.10
 - Add cooldown to Dependabot settings template
+- Do not allow double quotes in author name or package description
 
 ### Internal
 
 - Move log configuration to __init__.py
 - Set logging level via application configuration instead of hard-coding
 - Add cooldown to project Dependabot settings
-- Switching default quote style from single to double
+- Switch default quote style from single to double
 - Pin GitHub actions uv installs to the same version
 
 ## 0.5.7